shader terrain object

+++ c_lib/Necronomicon.c
@@ -103,7 +103,7 @@ float period_usecs = 0;
 const jack_time_t USECS_PER_SECOND = 1000000;
 double usecs_per_frame = 1000000 / 44100;
 
-double out_bus_buffers[16][512];
+float out_bus_buffers[16][512];
 unsigned int  out_bus_buffer_index = 0;
 
 /////////////////////
@@ -1546,14 +1546,14 @@ int process(jack_nframes_t nframes, void* arg)
 	double* _necronomicon_buses_out6 = _necronomicon_buses + BLOCK_SIZE * 6;
 	double* _necronomicon_buses_out7 = _necronomicon_buses + BLOCK_SIZE * 7;
 
-	double* _out_bus_0 = out_bus_buffers[0];
-	double* _out_bus_1 = out_bus_buffers[1];
-	double* _out_bus_2 = out_bus_buffers[2];
-	double* _out_bus_3 = out_bus_buffers[3];
-	double* _out_bus_4 = out_bus_buffers[4];
-	double* _out_bus_5 = out_bus_buffers[5];
-	double* _out_bus_6 = out_bus_buffers[6];
-	double* _out_bus_7 = out_bus_buffers[7];
+	float* _out_bus_0 = out_bus_buffers[0];
+	float* _out_bus_1 = out_bus_buffers[1];
+	float* _out_bus_2 = out_bus_buffers[2];
+	float* _out_bus_3 = out_bus_buffers[3];
+	float* _out_bus_4 = out_bus_buffers[4];
+	float* _out_bus_5 = out_bus_buffers[5];
+	float* _out_bus_6 = out_bus_buffers[6];
+	float* _out_bus_7 = out_bus_buffers[7];
 
 	unsigned int i;
 	for (i = 0; i < nframes; ++i)
@@ -5418,7 +5418,7 @@ float out_bus_rms(int bus)
 		return 0.0;
 
 	float   squareSum    = 0;
-	double* outBusBuffer = out_bus_buffers[bus];
+	float*  outBusBuffer = out_bus_buffers[bus];
 	uint    i            = 0;
 
 	for(i<512; ++i;)
